requirements.txt
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.5.0
django-debug-toolbar==3.2.1
django-extensions==3.1.1
django-filter==2.4.0
django-on-heroku==1.1.2
django-phonenumber-field==5.0.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
identify==2.2.11
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
nodeenv==1.6.0
packaging==21.0
phonenumbers==8.12.19
Pillow==8.1.2
platformdirs==2.2.0
pre-commit==2.13.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.6.0
whitenoise==5.3.0
